refactoring ui-code
